+++ NEWS.md
@@ -3,7 +3,7 @@
 * `map_dtr()`, `imap_dtr` and `pmap_dtr` now pass `.idcol` down to argument
   `idcol` of `data.table::rbindlist()`.
 * `cite_bib()` can now handle packages with multiple citation entries.
-
+* Add `msg_wrap()` and `cat_wrap()` for pretty printing of messages.
 
 # mlr3misc 0.1.6