+++ NEWS.md
@@ -3,7 +3,8 @@
 * `map_dtr()`, `imap_dtr` and `pmap_dtr` now pass `.idcol` down to argument
   `idcol` of `data.table::rbindlist()`.
 * `cite_bib()` can now handle packages with multiple citation entries.
-
+* Added argument `wrap` to `catf()`, `messagef()`, `warningf()` and `stopf()` to
+  wrap the respective messages to a desired length.
 
 # mlr3misc 0.1.6