+++ NEWS.md
@@ -1,3 +1,7 @@
+# mlr3misc 0.1.8
+
+* New function `check_packages_installed()`.
+
 # mlr3misc 0.1.7
 
 * `map_dtr()`, `imap_dtr()` and `pmap_dtr()` now pass `.idcol` down to argument